Move new helpers into blueprint/api
