+++ rerun_py/rerun_sdk/rerun/blueprint/__init__.py
@@ -1,5 +1,16 @@
 from __future__ import annotations
 
-__all__ = ["archetypes", "datatypes", "components"]
+__all__ = [
+    "archetypes",
+    "datatypes",
+    "components",
+    "Horizontal",
+    "Spatial2D",
+    "Spatial3D",
+    "Tabs",
+    "Vertical",
+    "Viewport",
+]
 
 from . import archetypes, components, datatypes
+from .api import Horizontal, Spatial2D, Spatial3D, Tabs, Vertical, Viewport
